Repository 병합 및 패키지 재배치, 엔티티 @setter 제거 등 기타 수정, ERD 수정

+++ backend/src/main/java/com/jobPrize/entity/common/Payment.java
@@ -38,7 +38,7 @@ public class Payment {
     private User user;
 	
 	@Column(name = "PAYMENT_AMOUNT", nullable = false)
-	private Long paymentAmount;
+	private int paymentAmount;
 	
 	@Column(name = "CONTENT", nullable = false)
 	private String content;
